requirements.txt
torch
torchvision
pyyaml
numpy
pillow
matplotlib
